requirements.txt
pandas
numpy
finpy-tse
matplotlib
arabic-reshaper
python-bidi
mplfinance
jdatetime
